requirements.txt
asgiref==3.3.1
Django==3.1.4
gunicorn==20.0.4
mysqlclient==2.0.1
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
